Default getTraceContextLater to getTraceContextAsync
